Add files via upload
This project was completed in Excel, we focused on the European banks data. With the provided information, we discovered the gender with the highest activity in the bank for each region, the average balances for each age group, the job classification that had the highest support for the bank.  The dash board created is fully interactive and can be manipulated by the end user for further insights.